+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.4.0] WIP
  - Allow limiting the depth of the recursion of `toJS()`
  - Allow disabling of the generic Buffer protocol transformation to `toJS()`
+ - Add `PyObject.prototype.map` method compatible with `Array.prototype.map`
 
 ### [1.3.1] WIP
  - Fix [#48](https://github.com/mmomtchev/pymport/issues/48), do not consider falsy properties as being undefined